Move futex syscall to its own file.
