+++ src/shims/posix/linux/mod.rs
@@ -1,2 +1,3 @@
 pub mod foreign_items;
 pub mod dlsym;
+pub mod sync;
